+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [5.2.0] - 2022-09-20
+### Added
+- Full support for multichain architecture
+
 ### Changed
 - Rename ForeignBlock model to BlockChainData
 - Execute migrations once before scheduling workers
@@ -14,16 +19,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Make jobs to run in parallel and make sure we have only one job per `chainId`
 
-## [5.1.3] - 2022-09-30
+### Removed
+- removed deprecated migrations
+
+## [5.1.3] - 2022-08-30
 ### Changed
 - Add logger when with synchronizer blocks length
 - Sync newest blocks first (debugging)
 
-## [5.1.2] - 2022-09-30
+## [5.1.2] - 2022-08-30
 ### Changed
 - Add logger when with synchronizer blocks length
 
-## [5.1.1] - 2022-09-30
+## [5.1.1] - 2022-08-30
 ### Changed
 - Revert sync newest blocks first